requirements.txt
streamlit
streamlit-drawable-canvas
pillow
pymupdf